refactor(aes): move general encryption methods to common
