+++ kipher-aes/src/main/kotlin/io/github/jhdcruz/kipher/aes/AesEncryption.kt
@@ -5,74 +5,27 @@
 
 package io.github.jhdcruz.kipher.aes
 
-import io.github.jhdcruz.kipher.common.KipherProvider
+import io.github.jhdcruz.kipher.common.KipherEncryption
 import org.jetbrains.annotations.NotNull
 import java.security.Provider
-import java.security.SecureRandom
 import javax.crypto.Cipher
-import javax.crypto.KeyGenerator
 
 // Constants
-internal const val ALGORITHM = "AES"
-internal const val AES_BLOCK_SIZE = 128
+internal const val ALGORITHM: String = "AES"
+internal const val AES_BLOCK_SIZE: Int = 128
 
 /**
  * Provides common functionalities for AES encryption.
  *
- * Should not be used/consumed directly.
+ * Should not be used/consumed directly. Unless for changing provider
  *
  * @param aesMode AES mode used for encryption.
  */
-sealed class AesEncryption(@NotNull aesMode: AesModes) : KipherProvider(provider) {
-    private val randomize = SecureRandom()
-    private val keyGenerator: KeyGenerator = KeyGenerator.getInstance(ALGORITHM)
+sealed class AesEncryption(@NotNull aesMode: AesModes) : KipherEncryption(ALGORITHM, provider) {
 
     internal val cipher: Cipher = Cipher.getInstance(aesMode.mode)
 
-    /**
-     * Generate a random IV based on [length].
-     */
-    internal fun generateIv(@NotNull length: Int): ByteArray {
-        return ByteArray(length).also {
-            randomize.nextBytes(it)
-        }
-    }
-
-    /**
-     * Generate a secret key.
-     *
-     * Reusing the same key for multiple encryption is not recommended,
-     * and poses security risk.
-     *
-     * `encrypt()` functions already generates a new key for each encryption.
-     */
-    fun generateKey(@NotNull keySize: Int = DEFAULT_KEY_SIZE): ByteArray {
-        return keyGenerator.run {
-            init(keySize, randomize)
-            generateKey().encoded
-        }
-    }
-
-    /**
-     * Concatenate the encryption details from a [Map] data.
-     *
-     * @return [ByteArray] Concatenated data
-     */
-    abstract fun Map<String, ByteArray>.concat(): ByteArray
-
-    /**
-     * Extracts the encryption details from the [ByteArray] data.
-     *
-     * Presumably encrypted using `encrypt()` functions
-     *
-     * @return [Map] of encryption details (such as iv, data, etc.)
-     */
-    abstract fun ByteArray.extract(): Map<String, ByteArray>
-
     companion object {
-        /** Default key size value. */
-        const val DEFAULT_KEY_SIZE: Int = 256
-
         /** Set JCE security provider. */
         var provider: Provider? = null
     }
